refactor(studio): extract useSnippetEditor from MonacoEditor
Move the snippet editing lifecycle out of MonacoEditor into a co-located
useSnippetEditor hook: tracking the editor text, creating the snippet in the
store on first edit (and routing to its URL, replacing vs pushing for a
?content= deep link), persisting debounced changes via setSql, and seeding the
editor from the content param.

MonacoEditor keeps only the editor shell and Monaco action wiring, consuming
{ snippet, disableEdit, handleEditorChange } from the hook. Behaviour-preserving.
